+++ src/bounds.rs
@@ -31,6 +31,7 @@ use crate::{EdgeDirection, Hex};
 /// [wraparound]: https://www.redblobgames.com/grids/hexagons/#wraparound
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "facet", derive(facet::Facet))]
 #[cfg_attr(feature = "bevy_reflect", derive(bevy_reflect::Reflect))]
 pub struct HexBounds {
     /// bounds center
@@ -93,6 +94,15 @@ impl HexBounds {
     /// Returns the number of hexagons in bounds
     pub const fn hex_count(&self) -> usize {
         Hex::range_count(self.radius) as usize
+    }
+
+    #[must_use]
+    #[inline]
+    #[doc(alias = "coords_count")]
+    #[doc(alias = "len32")]
+    /// Returns the number of hexagons in bounds
+    pub const fn hex_count32(&self) -> u32 {
+        Hex::range_count(self.radius)
     }
 
     #[doc(alias = "all_items")]